chore: reuse existing bucket creation modal for files
