+++ apps/studio/components/interfaces/Storage/FilesBuckets.tsx
@@ -2,9 +2,13 @@ import { useState } from 'react'
 
 import { useParams } from 'common'
 import { EmptyBucketState } from './EmptyBucketState'
-import { CreateBucketModalV2 } from './CreateBucketModalV2'
+import { CreateBucketModal } from './CreateBucketModal'
 import { useBucketsQuery } from 'data/storage/buckets-query'
 import { Bucket } from 'data/storage/buckets-query'
+import { ScaffoldSection } from 'components/layouts/Scaffold'
+import { FormHeader } from 'components/ui/Forms/FormHeader'
+import { Button } from 'ui'
+import { Plus } from 'lucide-react'
 
 export const FilesBuckets = () => {
   const { ref } = useParams()
@@ -31,45 +35,39 @@ export const FilesBuckets = () => {
     return (
       <>
         <EmptyBucketState bucketType="files" onCreateBucket={handleCreateBucket} />
-        <CreateBucketModalV2
-          visible={showCreateBucketModal}
-          onOpenChange={setShowCreateBucketModal}
-          defaultBucketType="STANDARD"
-        />
+        <CreateBucketModal hideAnalyticsOption={true} />
       </>
     )
   }
 
   // TODO: Implement bucket list view when buckets exist
   return (
-    <div className="space-y-4">
-      <div className="flex items-center justify-between">
-        <h2 className="text-lg font-medium text-foreground">Files Buckets</h2>
-        <button
-          onClick={handleCreateBucket}
-          className="px-4 py-2 bg-primary text-primary-foreground rounded-md hover:bg-primary/90 transition-colors"
-        >
-          Create Bucket
-        </button>
-      </div>
+    <div className="space-y-12">
+      <ScaffoldSection>
+        <div className="col-span-12">
+          <FormHeader
+            title="Buckets"
+            actions={
+              //   <Button type="primary" icon={<Plus size={14} />} onClick={handleCreateBucket}>
+              //     New media bucket
+              //   </Button>
+              <CreateBucketModal hideAnalyticsOption={true} />
+            }
+          />
 
-      <div className="grid gap-4">
-        {filesBuckets.map((bucket) => (
-          <div key={bucket.id} className="p-4 border rounded-lg">
-            <h3 className="font-medium text-foreground">{bucket.name}</h3>
-            <p className="text-sm text-foreground-light">
-              {bucket.public ? 'Public' : 'Private'} • Created{' '}
-              {new Date(bucket.created_at).toLocaleDateString()}
-            </p>
+          <div className="grid gap-4">
+            {filesBuckets.map((bucket) => (
+              <div key={bucket.id} className="p-4 border rounded-lg">
+                <h3 className="font-medium text-foreground">{bucket.name}</h3>
+                <p className="text-sm text-foreground-light">
+                  {bucket.public ? 'Public' : 'Private'} • Created{' '}
+                  {new Date(bucket.created_at).toLocaleDateString()}
+                </p>
+              </div>
+            ))}
           </div>
-        ))}
-      </div>
-
-      <CreateBucketModalV2
-        visible={showCreateBucketModal}
-        onOpenChange={setShowCreateBucketModal}
-        defaultBucketType="STANDARD"
-      />
+        </div>
+      </ScaffoldSection>
     </div>
   )
 }
